Finish structure and unit tests for command and response types
